+++ packages/internet-header/src/components.d.ts
@@ -5,38 +5,16 @@
  * It contains typing information for all components that exist in this project.
  */
 import { HTMLStencilElement, JSXBase } from "@stencil/core/internal";
-import { Link } from "./models/general.model";
+import { LinkConfig } from "./models/shared.model";
 import { ActiveRouteProp, Environment } from "./models/general.model";
-export { Link } from "./models/general.model";
+export { LinkConfig } from "./models/shared.model";
 export { ActiveRouteProp, Environment } from "./models/general.model";
 export namespace Components {
-    /**
-     * Trap the focus inside a specific container.
-     * @param active activate or deactivate the focus trap
-     */
-    interface FocusTrap {
-        /**
-          * @default false
-         */
-        "active": boolean;
-    }
-    interface PostKlpLoginWidget {
-        /**
-          * Override the logout-url provided by the portal config.
-         */
-        "logoutUrl"?: string;
-        /**
-          * Sets the focus on the login button
-         */
-        "setFocus": () => Promise<void>;
-    }
-    interface PostSkiplinks {
-    }
     interface SwisspostInternetBreadcrumbs {
         /**
           * Add custom breadcrumb items to the end of the pre-configured list. Handy if your online service has it's own navigation structure.
          */
-        "customItems"?: string | Array<Link>;
+        "customItems"?: string | Array<LinkConfig>;
         /**
           * Accessible label for the breadcrumbs navigation.
          */
@@ -123,28 +101,6 @@ export interface SwisspostInternetHeaderCustomEvent<T> extends CustomEvent<T> {
     target: HTMLSwisspostInternetHeaderElement;
 }
 declare global {
-    /**
-     * Trap the focus inside a specific container.
-     * @param active activate or deactivate the focus trap
-     */
-    interface HTMLFocusTrapElement extends Components.FocusTrap, HTMLStencilElement {
-    }
-    var HTMLFocusTrapElement: {
-        prototype: HTMLFocusTrapElement;
-        new (): HTMLFocusTrapElement;
-    };
-    interface HTMLPostKlpLoginWidgetElement extends Components.PostKlpLoginWidget, HTMLStencilElement {
-    }
-    var HTMLPostKlpLoginWidgetElement: {
-        prototype: HTMLPostKlpLoginWidgetElement;
-        new (): HTMLPostKlpLoginWidgetElement;
-    };
-    interface HTMLPostSkiplinksElement extends Components.PostSkiplinks, HTMLStencilElement {
-    }
-    var HTMLPostSkiplinksElement: {
-        prototype: HTMLPostSkiplinksElement;
-        new (): HTMLPostSkiplinksElement;
-    };
     interface HTMLSwisspostInternetBreadcrumbsElement extends Components.SwisspostInternetBreadcrumbs, HTMLStencilElement {
     }
     var HTMLSwisspostInternetBreadcrumbsElement: {
@@ -175,38 +131,19 @@ declare global {
         new (): HTMLSwisspostInternetHeaderElement;
     };
     interface HTMLElementTagNameMap {
-        "focus-trap": HTMLFocusTrapElement;
-        "post-klp-login-widget": HTMLPostKlpLoginWidgetElement;
-        "post-skiplinks": HTMLPostSkiplinksElement;
         "swisspost-internet-breadcrumbs": HTMLSwisspostInternetBreadcrumbsElement;
         "swisspost-internet-footer": HTMLSwisspostInternetFooterElement;
         "swisspost-internet-header": HTMLSwisspostInternetHeaderElement;
     }
 }
 declare namespace LocalJSX {
-    /**
-     * Trap the focus inside a specific container.
-     * @param active activate or deactivate the focus trap
-     */
-    interface FocusTrap {
-        /**
-          * @default false
-         */
-        "active"?: boolean;
-    }
-    interface PostKlpLoginWidget {
-        /**
-          * Override the logout-url provided by the portal config.
-         */
-        "logoutUrl"?: string;
-    }
-    interface PostSkiplinks {
-    }
+    type OneOf<K extends string, PropT, AttrT = PropT> = { [P in K]: PropT } & { [P in `attr:${K}` | `prop:${K}`]?: never } | { [P in `attr:${K}`]: AttrT } & { [P in K | `prop:${K}`]?: never } | { [P in `prop:${K}`]: PropT } & { [P in K | `attr:${K}`]?: never };
+
     interface SwisspostInternetBreadcrumbs {
         /**
           * Add custom breadcrumb items to the end of the pre-configured list. Handy if your online service has it's own navigation structure.
          */
-        "customItems"?: string | Array<Link>;
+        "customItems"?: string | Array<LinkConfig>;
         /**
           * Accessible label for the breadcrumbs navigation.
          */
@@ -291,29 +228,46 @@ declare namespace LocalJSX {
          */
         "textUserLinks": string;
     }
+
+    interface SwisspostInternetBreadcrumbsAttributes {
+        "customItems": string | Array<LinkConfig>;
+        "textHome": string;
+        "textBreadcrumbs": string;
+        "textMoreItems": string;
+    }
+    interface SwisspostInternetFooterAttributes {
+        "textFooter": string;
+        "textCookieSettings": string;
+    }
+    interface SwisspostInternetHeaderAttributes {
+        "activeRoute": ActiveRouteProp;
+        "environment": Environment;
+        "fullWidth": boolean;
+        "language": 'de' | 'fr' | 'it' | 'en';
+        "project": string;
+        "textBack": string;
+        "textChangeLanguage": string;
+        "textClose": string;
+        "textCurrentLanguage": string;
+        "textCurrentUser": string;
+        "textMain": string;
+        "textMenu": string;
+        "textUserLinks": string;
+    }
+
     interface IntrinsicElements {
-        "focus-trap": FocusTrap;
-        "post-klp-login-widget": PostKlpLoginWidget;
-        "post-skiplinks": PostSkiplinks;
-        "swisspost-internet-breadcrumbs": SwisspostInternetBreadcrumbs;
-        "swisspost-internet-footer": SwisspostInternetFooter;
-        "swisspost-internet-header": SwisspostInternetHeader;
+        "swisspost-internet-breadcrumbs": Omit<SwisspostInternetBreadcrumbs, keyof SwisspostInternetBreadcrumbsAttributes> & { [K in keyof SwisspostInternetBreadcrumbs & keyof SwisspostInternetBreadcrumbsAttributes]?: SwisspostInternetBreadcrumbs[K] } & { [K in keyof SwisspostInternetBreadcrumbs & keyof SwisspostInternetBreadcrumbsAttributes as `attr:${K}`]?: SwisspostInternetBreadcrumbsAttributes[K] } & { [K in keyof SwisspostInternetBreadcrumbs & keyof SwisspostInternetBreadcrumbsAttributes as `prop:${K}`]?: SwisspostInternetBreadcrumbs[K] } & OneOf<"textHome", SwisspostInternetBreadcrumbs["textHome"], SwisspostInternetBreadcrumbsAttributes["textHome"]> & OneOf<"textBreadcrumbs", SwisspostInternetBreadcrumbs["textBreadcrumbs"], SwisspostInternetBreadcrumbsAttributes["textBreadcrumbs"]> & OneOf<"textMoreItems", SwisspostInternetBreadcrumbs["textMoreItems"], SwisspostInternetBreadcrumbsAttributes["textMoreItems"]>;
+        "swisspost-internet-footer": Omit<SwisspostInternetFooter, keyof SwisspostInternetFooterAttributes> & { [K in keyof SwisspostInternetFooter & keyof SwisspostInternetFooterAttributes]?: SwisspostInternetFooter[K] } & { [K in keyof SwisspostInternetFooter & keyof SwisspostInternetFooterAttributes as `attr:${K}`]?: SwisspostInternetFooterAttributes[K] } & { [K in keyof SwisspostInternetFooter & keyof SwisspostInternetFooterAttributes as `prop:${K}`]?: SwisspostInternetFooter[K] } & OneOf<"textFooter", SwisspostInternetFooter["textFooter"], SwisspostInternetFooterAttributes["textFooter"]> & OneOf<"textCookieSettings", SwisspostInternetFooter["textCookieSettings"], SwisspostInternetFooterAttributes["textCookieSettings"]>;
+        "swisspost-internet-header": Omit<SwisspostInternetHeader, keyof SwisspostInternetHeaderAttributes> & { [K in keyof SwisspostInternetHeader & keyof SwisspostInternetHeaderAttributes]?: SwisspostInternetHeader[K] } & { [K in keyof SwisspostInternetHeader & keyof SwisspostInternetHeaderAttributes as `attr:${K}`]?: SwisspostInternetHeaderAttributes[K] } & { [K in keyof SwisspostInternetHeader & keyof SwisspostInternetHeaderAttributes as `prop:${K}`]?: SwisspostInternetHeader[K] } & OneOf<"project", SwisspostInternetHeader["project"], SwisspostInternetHeaderAttributes["project"]> & OneOf<"textBack", SwisspostInternetHeader["textBack"], SwisspostInternetHeaderAttributes["textBack"]> & OneOf<"textChangeLanguage", SwisspostInternetHeader["textChangeLanguage"], SwisspostInternetHeaderAttributes["textChangeLanguage"]> & OneOf<"textClose", SwisspostInternetHeader["textClose"], SwisspostInternetHeaderAttributes["textClose"]> & OneOf<"textCurrentLanguage", SwisspostInternetHeader["textCurrentLanguage"], SwisspostInternetHeaderAttributes["textCurrentLanguage"]> & OneOf<"textCurrentUser", SwisspostInternetHeader["textCurrentUser"], SwisspostInternetHeaderAttributes["textCurrentUser"]> & OneOf<"textMain", SwisspostInternetHeader["textMain"], SwisspostInternetHeaderAttributes["textMain"]> & OneOf<"textMenu", SwisspostInternetHeader["textMenu"], SwisspostInternetHeaderAttributes["textMenu"]> & OneOf<"textUserLinks", SwisspostInternetHeader["textUserLinks"], SwisspostInternetHeaderAttributes["textUserLinks"]>;
     }
 }
 export { LocalJSX as JSX };
 declare module "@stencil/core" {
     export namespace JSX {
         interface IntrinsicElements {
-            /**
-             * Trap the focus inside a specific container.
-             * @param active activate or deactivate the focus trap
-             */
-            "focus-trap": LocalJSX.FocusTrap & JSXBase.HTMLAttributes<HTMLFocusTrapElement>;
-            "post-klp-login-widget": LocalJSX.PostKlpLoginWidget & JSXBase.HTMLAttributes<HTMLPostKlpLoginWidgetElement>;
-            "post-skiplinks": LocalJSX.PostSkiplinks & JSXBase.HTMLAttributes<HTMLPostSkiplinksElement>;
-            "swisspost-internet-breadcrumbs": LocalJSX.SwisspostInternetBreadcrumbs & JSXBase.HTMLAttributes<HTMLSwisspostInternetBreadcrumbsElement>;
-            "swisspost-internet-footer": LocalJSX.SwisspostInternetFooter & JSXBase.HTMLAttributes<HTMLSwisspostInternetFooterElement>;
-            "swisspost-internet-header": LocalJSX.SwisspostInternetHeader & JSXBase.HTMLAttributes<HTMLSwisspostInternetHeaderElement>;
+            "swisspost-internet-breadcrumbs": LocalJSX.IntrinsicElements["swisspost-internet-breadcrumbs"] & JSXBase.HTMLAttributes<HTMLSwisspostInternetBreadcrumbsElement>;
+            "swisspost-internet-footer": LocalJSX.IntrinsicElements["swisspost-internet-footer"] & JSXBase.HTMLAttributes<HTMLSwisspostInternetFooterElement>;
+            "swisspost-internet-header": LocalJSX.IntrinsicElements["swisspost-internet-header"] & JSXBase.HTMLAttributes<HTMLSwisspostInternetHeaderElement>;
         }
     }
 }