+++ apps/integration-next/CHANGELOG.md
@@ -1,5 +1,13 @@
 # @swisspost/design-system-integration-next
 
+## 10.0.0-next.68
+
+### Patch Changes
+
+- Updated dependencies:
+  - @swisspost/design-system-styles@10.0.0-next.68
+  - @swisspost/design-system-components-react@10.0.0-next.68
+
 ## 10.0.0-next.67
 
 ### Major Changes